requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.15
    # via langchain-community
aiosignal==1.4.0
    # via aiohttp
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via httpx
attrs==25.3.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==6.1.0
    # via streamlit
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via streamlit
colorama==0.4.6
    # via
    #   click
    #   tqdm
dataclasses-json==0.6.7
    # via langchain-community
faiss-cpu==1.11.0.post1
    # via -r requirements.in
filelock==3.18.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.7.0
    # via
    #   huggingface-hub
    #   torch
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via streamlit
greenlet==3.2.3
    # via sqlalchemy
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via langsmith
httpx-sse==0.4.1
    # via langchain-community
huggingface-hub==0.34.3
    # via
    #   langchain-huggingface
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
    #   torch
joblib==1.5.1
    # via scikit-learn
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.25.0
    # via altair
jsonschema-specifications==2025.4.1
    # via jsonschema
langchain==0.3.27
    # via
    #   -r requirements.in
    #   langchain-community
langchain-community==0.3.27
    # via -r requirements.in
langchain-core==0.3.72
    # via
    #   langchain
    #   langchain-community
    #   langchain-huggingface
    #   langchain-text-splitters
langchain-huggingface==0.3.1
    # via -r requirements.in
langchain-text-splitters==0.3.9
    # via langchain
langsmith==0.4.13
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
markupsafe==3.0.2
    # via jinja2
marshmallow==3.26.1
    # via dataclasses-json
mpmath==1.3.0
    # via sympy
multidict==6.6.3
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via typing-inspect
narwhals==2.0.1
    # via altair
networkx==3.5
    # via torch
numpy==2.3.2
    # via
    #   -r requirements.in
    #   faiss-cpu
    #   langchain-community
    #   pandas
    #   pydeck
    #   scikit-learn
    #   scipy
    #   streamlit
    #   transformers
orjson==3.11.1
    # via langsmith
packaging==25.0
    # via
    #   altair
    #   faiss-cpu
    #   huggingface-hub
    #   langchain-core
    #   langsmith
    #   marshmallow
    #   streamlit
    #   transformers
pandas==2.3.1
    # via streamlit
pillow==11.3.0
    # via
    #   sentence-transformers
    #   streamlit
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
protobuf==6.31.1
    # via streamlit
pyarrow==21.0.0
    # via streamlit
pydantic==2.11.7
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-core
    #   langsmith
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via langchain-community
pydeck==0.9.1
    # via streamlit
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via
    #   -r requirements.in
    #   pydantic-settings
pytz==2025.2
    # via
    #   -r requirements.in
    #   pandas
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   langchain
    #   langchain-community
    #   langchain-core
    #   transformers
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2025.7.34
    # via transformers
requests==2.32.4
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   langchain
    #   langchain-community
    #   langsmith
    #   requests-toolbelt
    #   streamlit
    #   transformers
requests-toolbelt==1.0.0
    # via langsmith
rpds-py==0.26.0
    # via
    #   jsonschema
    #   referencing
safetensors==0.6.1
    # via
    #   -r requirements.in
    #   transformers
scikit-learn==1.7.1
    # via sentence-transformers
scipy==1.16.1
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==5.1.0
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.42
    # via
    #   langchain
    #   langchain-community
streamlit==1.48.0
    # via -r requirements.in
sympy==1.14.0
    # via torch
tenacity==9.1.2
    # via
    #   langchain-community
    #   langchain-core
    #   streamlit
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.21.4
    # via
    #   langchain-huggingface
    #   transformers
toml==0.10.2
    # via streamlit
torch==2.8.0
    # via
    #   -r requirements.in
    #   sentence-transformers
tornado==6.5.1
    # via streamlit
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   sentence-transformers
    #   transformers
transformers==4.55.0
    # via
    #   -r requirements.in
    #   sentence-transformers
typing-extensions==4.14.1
    # via
    #   aiosignal
    #   altair
    #   anyio
    #   huggingface-hub
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   referencing
    #   sentence-transformers
    #   sqlalchemy
    #   streamlit
    #   torch
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
watchdog==6.0.0
    # via streamlit
yarl==1.20.1
    # via aiohttp
zstandard==0.23.0
    # via langsmith
